refactor: 홈화면 리팩토링
