+++ app/src/main/java/com/apptive/japkor/data/model/Matching.kt
@@ -29,3 +29,15 @@ data class MalePendingMatchingResponse(
     @SerializedName(value = "profileImageUrls", alternate = ["profileImageUrl"])
     val profileImageUrls: List<String>?
 )
+
+data class HomeMatching(
+    val matchingId: Long,
+    val memberId: Long,
+    val name: String,
+    val email: String,
+    val height: Int?,
+    val weight: Int?,
+    val residenceArea: String?,
+    val matchingOrder: Int?,
+    val status: String
+)
