--- app/src/main/java/com/apptive/japkor/ui/main/home/HomeMatching.kt
@@ -1,13 +1,0 @@
-package com.apptive.japkor.ui.main.home
-
-data class HomeMatching(
-    val matchingId: Long,
-    val memberId: Long,
-    val name: String,
-    val email: String,
-    val height: Int?,
-    val weight: Int?,
-    val residenceArea: String?,
-    val matchingOrder: Int?,
-    val status: String
-)
